Added Window, added WorldClock, toggle teleport (keyboard T), pickup item
added Matt's MyClock files under the name WorldClock. Created a Window blueprint, made some changes to teleporting (not sure if we will end up using this or not) so it can be toggled on and off by pressing T.  added two objects into the room to test picking up and dropping items. picking up and dropping is done be clicking when looking at an item that simulates physics right now.